+++ mapscript/php/querymap.c
@@ -57,7 +57,7 @@ PHP_METHOD(queryMapObj, __construct)
 PHP_METHOD(queryMapObj, __get)
 {
     char *property;
-    int property_len;
+    long property_len;
     zval *zobj = getThis();
     php_querymap_object *php_querymap;
 
@@ -84,7 +84,7 @@ PHP_METHOD(queryMapObj, __get)
 PHP_METHOD(queryMapObj, __set)
 {
     char *property;
-    int property_len;
+    long property_len;
     zval *value;
     zval *zobj = getThis();
     php_querymap_object *php_querymap;
@@ -117,7 +117,7 @@ PHP_METHOD(queryMapObj, __set)
 PHP_METHOD(queryMapObj, updateFromString)
 {
     char *snippet;
-    int snippet_len;
+    long snippet_len;
     zval *zobj = getThis();
     php_querymap_object *php_querymap;
     int status = MS_FAILURE;
